requirements.txt
django
psycopg2-binary
Pillow
django-quill-editor
django-colorfield